+++ pages/_app.js
@@ -131,7 +131,7 @@ function MyApp(props) {
                   logoProjeto: width > 1000 ?
                     path == '/' ? "https://media.graphassets.com/3Vvlszx1RraNWFWyfgaT" : props.res[0].logoIps[0].logo[0].url :
                     props.res[0].logoIps[1].logo[0].url,
-                  cor: (path == '/' || path == '/apoio' || path == '/analise' || path == '/impulsogov') ? "Cinza" : "White",
+                  cor: (path == '/' || path == '/apoio' || path == '/analise' || path == 'quemsomos') ? "Cinza" : "White",
                   logoLink: props.ses ? '/inicio' : '/'
                 } }
